requirements.txt
async-timeout==4.0.3
ble==0
bleak==0.20.2
bluepy==1.3.0
dbus-fast==1.91.2
paho-mqtt==1.6.1
